requirements.txt
llama-index==0.10.0
llama-index-embeddings-huggingface==0.2.0
llama-index-llms-openai==0.1.0
llama-index-readers-file==0.1.0
openai==1.12.0
sentence-transformers==2.3.1
torch==2.1.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pypdf==3.17.0
werkzeug==3.0.1
